Moved UserProfile, ChatInfo, ChatForm to ui-kit components (#189)
* migrated ConversationItem & ConversationItemList to components

* Sync tailwind version, corrected css style for components

* replaced ChatInfo, EditModals & ChatInfo to ui-kit

* added prettier, eslint and tailwind configs

* Updated configs, migrated ChatForm components (#190)

* init of additional configs, cleaning, changing tailwindcss to prefix format

* migrated ChatFromHeader, added useContextMenu, useMessages

* replaced ChatFromInput

* minor style changes, fixed some non-working input field functions

* updated DynamicAvatar component, global styles, ChatMessage structure

* updated ChatMessage structure

* migrated useConfirmWindow

* fixed height of ChatList container

+++ apps/client/src/components/context/elements/ConversationActions.js
@@ -4,28 +4,23 @@ import { useSelector } from "react-redux";
 
 import conversationService from "@services/conversationsService";
 
+import { useConfirmWindow } from "@sama-communications.ui-kit";
+
 import ContextLink from "@components/context/elements/ContextLink";
 
 import { selectContextExternalProps } from "@store/values/ContextMenu.js";
 import { getIsTabletView } from "@store/values/IsTabletView.js";
-import {
-  getConverastionById,
-  selectAllConversations,
-} from "@store/values/Conversations.js";
+import { getConverastionById, selectAllConversations } from "@store/values/Conversations.js";
 
 import { addPrefix, addSuffix, navigateTo } from "@utils/NavigationUtils.js";
 
-import Info from "@icons/context/Info.svg?react";
-import Edit from "@icons/context/Edit.svg?react";
-import Leave from "@icons/context/Leave.svg?react";
-import Add from "@icons/context/AddParticipants.svg?react";
-import Remove from "@icons/context/Remove.svg?react";
-import SendMessage from "@icons/context/SendMessage.svg?react";
+import { Info, MessageCircleOff, MessageSquareMore, SquarePen, Trash, UserMinus, UserPlus } from "lucide-react";
 
 export default function ConversationActions({ listOfIds }) {
+  const confirm = useConfirmWindow();
+
   const { pathname, hash } = useLocation();
-  const { type, opponent_id, owner_id } =
-    useSelector(getConverastionById) || {};
+  const { type, opponent_id, owner_id } = useSelector(getConverastionById) || {};
 
   const currentUser = useSelector(selectAllConversations);
   const currentPath = pathname + hash;
@@ -39,16 +34,11 @@ export default function ConversationActions({ listOfIds }) {
       <ContextLink
         key={"convInfo"}
         text="Info"
-        icon={<Info />}
+        icon={<Info size={18} />}
         onClick={() => {
           const tmpPath =
-            isTabletView && currentPath.includes("/profile")
-              ? currentPath.replace("/profile", "")
-              : currentPath;
-          addSuffix(
-            tmpPath,
-            type === "g" ? "/info" : `/user?uid=${opponent_id}`
-          );
+            isTabletView && currentPath.includes("/profile") ? currentPath.replace("/profile", "") : currentPath;
+          addSuffix(tmpPath, type === "g" ? "/info" : `/user?uid=${opponent_id}`);
         }}
       />
     ),
@@ -56,7 +46,7 @@ export default function ConversationActions({ listOfIds }) {
       <ContextLink
         key={"convEdit"}
         text="Edit"
-        icon={<Edit />}
+        icon={<SquarePen size={18} />}
         onClick={() => addSuffix(currentPath, "/edit?type=chat")}
       />
     ),
@@ -64,9 +54,15 @@ export default function ConversationActions({ listOfIds }) {
       <ContextLink
         key={"convLeave"}
         text="Delete and leave"
-        icon={<Leave />}
+        icon={<Trash size={18} color="red" />}
         isDangerStyle={true}
         onClick={async () => {
+          const { isConfirm } = await confirm({
+            title: "Delate And Leave",
+            description: `Do you want to delete this chat?`,
+            icon: <MessageCircleOff size={40} color="red" strokeWidth={2} />,
+          });
+          if (!isConfirm) return;
           navigateTo("/");
           await conversationService.deleteConversation();
         }}
@@ -76,7 +72,7 @@ export default function ConversationActions({ listOfIds }) {
       <ContextLink
         key={"convAddParticipants"}
         text="Add participants"
-        icon={<Add />}
+        icon={<UserPlus size={18} />}
         onClick={() => addSuffix(currentPath, "/add")}
       />
     ),
@@ -84,7 +80,7 @@ export default function ConversationActions({ listOfIds }) {
       <ContextLink
         key={"convRemoveParticipants"}
         text="Remove participant"
-        icon={<Remove />}
+        icon={<UserMinus size={18} color="red" />}
         isDangerStyle={true}
         onClick={() => conversationService.removeParticipant(userObject?._id)}
       />
@@ -94,7 +90,7 @@ export default function ConversationActions({ listOfIds }) {
       <ContextLink
         key={"participantInfo"}
         text="Info"
-        icon={<Info />}
+        icon={<Info size={18} />}
         uId={userObject?._id}
         onClick={() => {
           isCurrentUserOwner
@@ -107,21 +103,15 @@ export default function ConversationActions({ listOfIds }) {
       <ContextLink
         key={"participantSendMessage"}
         text="Write a message"
-        icon={<SendMessage />}
+        icon={<MessageSquareMore size={18} />}
         uObject={userObject}
         onClick={async () => {
-          const chatId = await conversationService.createPrivateChat(
-            userObject?._id,
-            userObject
-          );
+          const chatId = await conversationService.createPrivateChat(userObject?._id, userObject);
           navigateTo(`/#${chatId}`);
         }}
       />
     ),
   };
 
-  return useMemo(
-    () => listOfIds.map((linkId) => links[linkId]).filter(Boolean),
-    [listOfIds, userObject]
-  );
+  return useMemo(() => listOfIds.map((linkId) => links[linkId]).filter(Boolean), [listOfIds, userObject]);
 }
